+++ examples/evolve_nqueens.rs
@@ -51,9 +51,8 @@ fn main() {
         .with_target_fitness_score(0)
         // .with_replace_on_equal_fitness(true) // not crucial for this problem
         .with_mutate(MutateSingleGene::new(0.2))
-        .with_crossover(CrossoverRejuvenate::new(0.9))
-        // .with_crossover(CrossoverClone::new(0.9))
-        .with_select(SelectElite::new(0.5, 0.5))
+        .with_crossover(CrossoverClone::new(0.7))
+        .with_select(SelectElite::new(0.5, 0.05))
         .with_reporter(EvolveReporterSimple::new(100))
         .build()
         .unwrap();